Update game-211820.sh
Reverted the hardcoded -v flag in mv command. The problem with mv not running was the result of a bug in the main script. The EXTENSION_AFTER callback was receiving the wrong parameters due to an undefined DEBUG variable, resulting in the parameters being EXTENSION_AFTER "user" "wsitems" when not run in debug mode. This has been fixed in SWDS 2.5.9